refactor(doctor): reduce duplication in files check
Move common code to files.go (no build tag) and only split out the
platform-specific checkDiskSpace function into files_unix.go and
files_windows.go.
